add link command
